requirements.txt
tornado
pycurl
bs4
lxml
motor
langdetect
python-dateutil
